Move code into separate files for clarity
